Squashed commit of the following:
commit ff3e4f83f56c4eee8e3889374cc502185c961afa
Merge: ff9d5683 aed19c26
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Feb 8 20:13:20 2023 +0100

    Merge branch 'master' into dev

commit ff9d5683b47f1fa4c5391e36735e15b4e82465bc
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Feb 8 20:11:07 2023 +0100

    Updated version.

commit 4143bb6095fb316ea0188b9c985b722d4fa02b15
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Feb 8 18:35:23 2023 +0100

    Tuned score only culling.

commit d1cd241fcb0ba622c1f5b1b1ddb3e0028760d77e
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Feb 8 11:50:41 2023 +0100

    Removed kmer filter.

commit d63cc38216fba58687593f72830f7545b2ee6e21
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Feb 8 11:30:27 2023 +0100

    Fixed clang errors.

commit abcec00aefce0fc42cd6b1cf9f29b504a0a3b824
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Tue Feb 7 20:50:55 2023 +0100

    Fixed DAA errors.

commit b76ab78f94f7fcb55bb9a43c42aaf2909800304b
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Tue Feb 7 17:32:06 2023 +0100

    Fixed recluster issues.

commit 5b6896da00435b46dfa349cac7309d2b2d55f702
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Tue Feb 7 17:15:08 2023 +0100

    Fixed warnings.

commit b45676d30d7ec00f193bcde7348cfa91240c2685
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Tue Feb 7 12:48:53 2023 +0100

    Fixed warnings.

commit 53d8b4ac96c671afc22730904023b34651983d7a
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Tue Feb 7 12:25:52 2023 +0100

    Fixed include.

commit 27bc731ac68e2af950b4b479d4fb0e462a8d4a7a
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Sun Feb 5 12:23:32 2023 +0100

    Fixed compilation errors.

commit a89ba6dac5383db6b0b0eb677d1e869e7383bc3a
Merge: ac6df028 8f119f1f
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Fri Feb 3 18:34:31 2023 +0100

    Conflicts

+++ src/align/global_ranking/table.cpp
@@ -181,7 +181,7 @@ void update_table(Search::Config& cfg) {
 		}
 	};
 	vector<thread> threads;
-	for (size_t i = 0; i < config.threads_; ++i)
+	for (int i = 0; i < config.threads_; ++i)
 		threads.emplace_back(worker);
 	for (thread& t : threads)
 		t.join();
